requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.11.1
icalendar==4.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
metapy==0.2.10
pkg-resources==0.0.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
tqdm==4.23.2
urllib3==1.22
Werkzeug==0.14.1
